Refactor InputView API and implementation
This commit rewrites a major part of the InputView API and
implementation. This class is now intended to be a front-end widget
rather than a complete input system. The new API now provides access to
the user-drawn glyphs as rasterized strokes in a form of floating point
buffers with user-requested resolution. The constructor now takes user
settings as optional parameters. The implementation merges multiple
strokes into one glyph if they are close enough.